+++ ddtrace/monkey.py
@@ -58,6 +58,7 @@ PATCH_MODULES = {
     "mako": True,
     "flask": True,
     "kombu": False,
+    "starlette": True,
     # Ignore some web framework integrations that might be configured explicitly in code
     "falcon": False,
     "pylons": False,
@@ -65,6 +66,7 @@ PATCH_MODULES = {
     # Auto-enable logging if the environment variable DD_LOGS_INJECTION is true
     "logging": config.logs_injection,
     "pynamodb": True,
+    "pyodbc": True,
 }
 
 _LOCK = threading.Lock()
@@ -84,6 +86,7 @@ _PATCH_ON_IMPORT = {
     "requests": ("requests",),
     "botocore": ("botocore",),
     "elasticsearch": ("elasticsearch",),
+    "pynamodb": ("pynamodb",),
 }
 
 
@@ -166,7 +169,10 @@ def patch(raise_errors=True, **patch_modules):
 
     patched_modules = get_patched_modules()
     log.info(
-        "patched %s/%s modules (%s)", len(patched_modules), len(modules), ",".join(patched_modules),
+        "patched %s/%s modules (%s)",
+        len(patched_modules),
+        len(modules),
+        ",".join(patched_modules),
     )