+++ ddtrace/profiling/profiler.py
@@ -3,11 +3,14 @@ import atexit
 import logging
 import os
 
+import ddtrace
 from ddtrace.profiling import recorder
 from ddtrace.profiling import scheduler
 from ddtrace.utils import deprecation
+from ddtrace.utils import formats
 from ddtrace.vendor import attr
 from ddtrace.profiling.collector import exceptions
+from ddtrace.profiling.collector import memalloc
 from ddtrace.profiling.collector import memory
 from ddtrace.profiling.collector import stack
 from ddtrace.profiling.collector import threading
@@ -72,7 +75,7 @@ class Profiler(object):
     service = attr.ib(factory=_get_service_name)
     env = attr.ib(factory=lambda: os.environ.get("DD_ENV"))
     version = attr.ib(factory=lambda: os.environ.get("DD_VERSION"))
-    tracer = attr.ib(default=None)
+    tracer = attr.ib(default=ddtrace.tracer)
     _collectors = attr.ib(init=False, default=None)
     _scheduler = attr.ib(init=False, default=None)
     status = attr.ib(init=False, type=ProfilerStatus, default=ProfilerStatus.STOPPED)
@@ -107,13 +110,20 @@ class Profiler(object):
                 # This can generate one event every 0.1s if 100% are taken — though we take 5% by default.
                 # = (60 seconds / 0.1 seconds)
                 memory.MemorySampleEvent: int(60 / 0.1),
+                # (default buffer size / interval) * export interval
+                memalloc.MemoryAllocSampleEvent: int((64 / 0.5) * 60),
             },
             default_max_events=int(os.environ.get("DD_PROFILING_MAX_EVENTS", recorder.Recorder._DEFAULT_MAX_EVENTS)),
         )
 
+        if formats.asbool(os.environ.get("DD_PROFILING_MEMALLOC", "false")):
+            mem_collector = memalloc.MemoryCollector(r)
+        else:
+            mem_collector = memory.MemoryCollector(r)
+
         self._collectors = [
             stack.StackCollector(r, tracer=self.tracer),
-            memory.MemoryCollector(r),
+            mem_collector,
             exceptions.UncaughtExceptionCollector(r),
             threading.LockCollector(r),
         ]
@@ -152,17 +162,17 @@ class Profiler(object):
 
         :param flush: Wait for the flush of the remaining events before stopping.
         """
+        if self._scheduler:
+            self._scheduler.stop()
+
         for col in reversed(self._collectors):
             col.stop()
 
         for col in reversed(self._collectors):
             col.join()
 
-        if self._scheduler:
-            self._scheduler.stop()
-
-            if flush:
-                self._scheduler.join()
+        if self._scheduler and flush:
+            self._scheduler.join()
 
         self.status = ProfilerStatus.STOPPED