+++ tests/contrib/django/test_django.py
@@ -1,3 +1,4 @@
+import itertools
 import django
 from django.views.generic import TemplateView
 from django.test import modify_settings, override_settings
@@ -5,8 +6,10 @@ import os
 import pytest
 
 from ddtrace import config
+from ddtrace.compat import string_type, binary_type
 from ddtrace.constants import ANALYTICS_SAMPLE_RATE_KEY, SAMPLING_PRIORITY_KEY
 from ddtrace.contrib.django.patch import instrument_view
+from ddtrace.contrib.django.utils import get_request_uri
 from ddtrace.ext import http, errors
 from ddtrace.ext.priority import USER_KEEP
 from ddtrace.propagation.http import HTTP_HEADER_TRACE_ID, HTTP_HEADER_PARENT_ID, HTTP_HEADER_SAMPLING_PRIORITY
@@ -189,7 +192,10 @@ def test_v2XX_middleware(client, test_spans):
     # Assert common span structure
     for span in middleware_spans:
         span.assert_matches(
-            name="django.middleware", service="django", error=0, span_type=None,
+            name="django.middleware",
+            service="django",
+            error=0,
+            span_type=None,
         )
 
     span_resources = {
@@ -286,7 +292,9 @@ def test_django_request_not_found(client, test_spans):
     render_span.assert_matches(
         name="django.template.render",
         resource="django.template.base.Template.render",
-        meta={"django.template.engine.class": "django.template.engine.Engine",},
+        meta={
+            "django.template.engine.class": "django.template.engine.Engine",
+        },
     )
 
 
@@ -395,7 +403,10 @@ def test_request_view(client, test_spans):
     # Assert span properties
     view_span = view_spans[0]
     view_span.assert_matches(
-        name="django.view", service="django", resource="tests.contrib.django.views.index", error=0,
+        name="django.view",
+        service="django",
+        resource="tests.contrib.django.views.index",
+        error=0,
     )
 
 
@@ -423,7 +434,10 @@ def test_template_view(client, test_spans):
     # Assert span properties
     view_span = view_spans[0]
     view_span.assert_matches(
-        name="django.view", service="django", resource="tests.contrib.django.views.template_view", error=0,
+        name="django.view",
+        service="django",
+        resource="tests.contrib.django.views.template_view",
+        error=0,
     )
 
     root_span = test_spans.get_root_span()
@@ -441,7 +455,10 @@ def test_template_simple_view(client, test_spans):
     # Assert span properties
     view_span = view_spans[0]
     view_span.assert_matches(
-        name="django.view", service="django", resource="tests.contrib.django.views.template_simple_view", error=0,
+        name="django.view",
+        service="django",
+        resource="tests.contrib.django.views.template_simple_view",
+        error=0,
     )
 
     root_span = test_spans.get_root_span()
@@ -459,7 +476,10 @@ def test_template_list_view(client, test_spans):
     # Assert span properties
     view_span = view_spans[0]
     view_span.assert_matches(
-        name="django.view", service="django", resource="tests.contrib.django.views.template_list_view", error=0,
+        name="django.view",
+        service="django",
+        resource="tests.contrib.django.views.template_list_view",
+        error=0,
     )
 
     root_span = test_spans.get_root_span()
@@ -541,7 +561,8 @@ def test_simple_view_get(client, test_spans):
     assert len(spans) == 1
     span = spans[0]
     span.assert_matches(
-        resource="tests.contrib.django.views.BasicView.get", error=0,
+        resource="tests.contrib.django.views.BasicView.get",
+        error=0,
     )
 
 
@@ -555,7 +576,8 @@ def test_simple_view_post(client, test_spans):
     assert len(spans) == 1
     span = spans[0]
     span.assert_matches(
-        resource="tests.contrib.django.views.BasicView.post", error=0,
+        resource="tests.contrib.django.views.BasicView.post",
+        error=0,
     )
 
 
@@ -568,7 +590,8 @@ def test_simple_view_delete(client, test_spans):
     assert len(spans) == 1
     span = spans[0]
     span.assert_matches(
-        resource="tests.contrib.django.views.BasicView.delete", error=0,
+        resource="tests.contrib.django.views.BasicView.delete",
+        error=0,
     )
 
 
@@ -582,7 +605,8 @@ def test_simple_view_options(client, test_spans):
     assert len(spans) == 1
     span = spans[0]
     span.assert_matches(
-        resource="django.views.generic.base.View.options", error=0,
+        resource="django.views.generic.base.View.options",
+        error=0,
     )
 
 
@@ -595,7 +619,8 @@ def test_simple_view_head(client, test_spans):
     assert len(spans) == 1
     span = spans[0]
     span.assert_matches(
-        resource="tests.contrib.django.views.BasicView.head", error=0,
+        resource="tests.contrib.django.views.BasicView.head",
+        error=0,
     )
 
 
@@ -1135,7 +1160,12 @@ def test_django_request_distributed(client, test_spans):
     # DEV: Do not use `test_spans.get_root_span()` since that expects `parent_id is None`
     root = test_spans.find_span(name="django.request")
     root.assert_matches(
-        name="django.request", trace_id=12345, parent_id=78910, metrics={SAMPLING_PRIORITY_KEY: USER_KEEP,},
+        name="django.request",
+        trace_id=12345,
+        parent_id=78910,
+        metrics={
+            SAMPLING_PRIORITY_KEY: USER_KEEP,
+        },
     )
 
 
@@ -1460,3 +1490,30 @@ def test_template_view_patching():
     # Patch via `as_view()`
     TemplateView.as_view()
     assert "dispatch" not in vars(TemplateView)
+
+
+class _HttpRequest(django.http.HttpRequest):
+    @property
+    def scheme(self):
+        return b"http"
+
+
+@pytest.mark.parametrize(
+    "request_cls,request_path,http_host",
+    itertools.product(
+        [django.http.HttpRequest, _HttpRequest],
+        [u"/;some/?awful/=path/foo:bar/", b"/;some/?awful/=path/foo:bar/"],
+        [u"testserver", b"testserver"],
+    ),
+)
+def test_helper_get_request_uri(request_cls, request_path, http_host):
+    request = request_cls()
+    request.path = request_path
+    request.META = {"HTTP_HOST": http_host}
+    request_uri = get_request_uri(request)
+    assert (
+        request_cls == _HttpRequest
+        and isinstance(request_path, binary_type)
+        and isinstance(http_host, binary_type)
+        and isinstance(request_uri, binary_type)
+    ) or isinstance(request_uri, string_type)