+++ tests/contrib/elasticsearch/test.py
@@ -44,7 +44,10 @@ class ElasticsearchTest(TracerTestCase):
         """
         tracer = get_dummy_tracer()
         writer = tracer.writer
-        transport_class = get_traced_transport(datadog_tracer=tracer, datadog_service=self.TEST_SERVICE,)
+        transport_class = get_traced_transport(
+            datadog_tracer=tracer,
+            datadog_service=self.TEST_SERVICE,
+        )
 
         es = elasticsearch.Elasticsearch(transport_class=transport_class, port=ELASTICSEARCH_CONFIG["port"])
 
@@ -154,7 +157,10 @@ class ElasticsearchTest(TracerTestCase):
         writer = tracer.writer
         ot_tracer = init_tracer("my_svc", tracer)
 
-        transport_class = get_traced_transport(datadog_tracer=tracer, datadog_service=self.TEST_SERVICE,)
+        transport_class = get_traced_transport(
+            datadog_tracer=tracer,
+            datadog_service=self.TEST_SERVICE,
+        )
 
         es = elasticsearch.Elasticsearch(transport_class=transport_class, port=ELASTICSEARCH_CONFIG["port"])