+++ CHANGELOG.md
@@ -1,11 +1,21 @@
 # Changelog
 
+---
+
+## 0.43.0 (5/10/2020)
+
+- fix(django): avoid mixing str and non-str args for uri helper
+- fix(asgi): tag 500-level responses as errors
+- fix(asgi): set http status when exception raised
+- fix(rediscluster): support rediscluster==2.1.0
 - fix(asyncio): enable patch by default
 - fix(asyncio): patch base event loop class
 - fix(vertica): use strings in `__all__`
 - feat(core): backport contextvars
 - fix(sanic): fix patching for sanic async http server (#1659)
 - fix(flask): make template patching idempotent
+- fix(core): Do not rate limit log lines when in debug
+- fix(profiling): Fix a potential deadlock on profiler restart after fork()
 
 ---